refactor(packages): name a file after what is in it
Four packages were shaped like this:

    packages/validator/index.js              49 lines — a re-export barrel
    packages/validator/internal/schema.js   767 lines — the entire library

The real code hid one directory down in a file named after nothing in
particular, and `index.js` was pure indirection. A reader opening the package
learnt nothing from its shape.

Judged one at a time, because the answer is not the same for all four:

**validator** and **effect** are each one concept. A schema is a closure, and
the primitives, combinators, `pipe` steps and two runners are four views of
it; `Effect`, `Fiber`, `Tag`, `Layer` and `Scope` are opaque types over
carriers one module defines, and splitting them would hand `makeEffect` and
`readKernel` to siblings — giving away the opacity inside the package to buy
a directory listing. Both are their `index.js` now, and `internal/` is gone.

**web** and **hooks** are not. Five different mistakes (a layout shift, a
hydration mismatch, an unannounced region, a cookie with two readers, a head
written too late) and six different subjects are six files, at the package
root, named after their domain, with the directory listing as the table of
contents. `internal/element.js` is `dom.js`: it was named after its argument
rather than its job, which made `browser.js` look like its opposite when it
is its sibling — `dom.js` takes a ref because there are as many answers as
elements, `browser.js` needs none because there is one browser.

`internal/` survives in exactly two places, `core/internal/native-runtime.js`
and `host/internal/node-hooks.js`, where it means "a consumer must not reach
this" and a sibling package genuinely does. That is what the directory is
for.

No public specifier changed. Eleven subpaths were *added*, because a module
outside `internal/` has to be reachable through `exports` and
`every_shipped_module_is_reachable_through_exports` says so.

Every moved file keeps its prose, and each root module gained a section
naming its boundary and the neighbour it is confused with — `Font` stays in
`media.js` though it is a `<link>`, because the hard part is the fetch.
