+++ packages/hooks/async.js
@@ -1,6 +1,6 @@
 // @flow
 //
-// Running a promise from a component.
+// `@uniflowed/hooks/async`: running a promise from a component.
 //
 // Two bugs a hand-written version has, and only one of them is a warning:
 // setting state after the component has gone, and a slow first request
@@ -12,6 +12,19 @@
 // because React runs its cleanup before running it again. That is the shape
 // React's own documentation uses, and it means there is no "latest" anything
 // to keep in a ref and no generation counter to keep in step.
+//
+// # What belongs in this module
+//
+// A hook that starts one call and holds its pending, resolved and failed
+// states. One file for one hook, because the subject is neither the
+// component's life nor a timer nor a DOM node, and hiding it inside one of
+// those would make all three harder to name.
+//
+// Deliberately not here, and not in this package at all: caching, retries,
+// deduplication between components, invalidation. Those are
+// `@uniflowed/query`'s, and the boundary is what keeps `useAsync` small enough
+// to read in one sitting. When a caller needs a cache they should change
+// packages, not discover that this hook grew one.
 
 import { useCallback, useEffect, useState } from "@uniflowed/react";
