+++ packages/hooks/index.js
@@ -20,10 +20,39 @@
 // Where there is no honest default, the caller supplies one: a page that hides
 // its sidebar under 48rem wants `false` on the server and one that renders a
 // mobile menu wants `true`, and a library cannot know which.
+//
+// # How the package is laid out
+//
+// Six modules beside this one, split by what a hook's subject is — because
+// that is the question a reader looking for one actually asks:
+//
+// - `lifecycle.js` — the component itself: mounted, previous, run once.
+// - `state.js` — a value the component owns, with the operations that suit it.
+// - `timing.js` — when something runs: intervals, timeouts, debounce,
+//   throttle.
+// - `async.js` — one promise, and the stale-response bug.
+// - `browser.js` — the ambient environment: viewport, connection, preferences.
+// - `dom.js` — one element the caller holds a ref to: listen, measure,
+//   observe.
+//
+// The two that are easiest to confuse are the last two, so each says so in its
+// own header: `browser.js` needs no ref because there is one browser, and
+// `dom.js` needs one because there are as many answers as there are elements.
+//
+// They sit here rather than under an `internal/`, and each has its own
+// subpath. Every name in them is exported from this file, so calling them
+// internal would have described nothing true, and it cost a reader a directory
+// hop to reach the first line of code. `internal/` is for a module consumers
+// must not reach; this package has none.
+//
+// `lifecycle.js` is the one the others import — `timing.js`, `state.js` and
+// `dom.js` all want `useStableCallback` — and it is still a subject rather
+// than a bag of shared helpers. A hook goes there because it is about the
+// component's life, not because more than one file wanted it.
 
-export type { Async } from "./internal/async.js";
+export type { Async } from "./async.js";
 
-export { useAsync } from "./internal/async.js";
+export { useAsync } from "./async.js";
 export {
   useIsomorphicLayoutEffect,
   useMount,
@@ -32,14 +61,14 @@ export {
   useRerender,
   useStableCallback,
   useUnmount,
-} from "./internal/lifecycle.js";
+} from "./lifecycle.js";
 export {
   useDebouncedCallback,
   useDebouncedValue,
   useInterval,
   useThrottledCallback,
   useTimeout,
-} from "./internal/timing.js";
+} from "./timing.js";
 export {
   useDocumentVisible,
   useMediaQuery,
@@ -47,7 +76,7 @@ export {
   usePrefersReducedMotion,
   usePreferredColorScheme,
   useWindowSize,
-} from "./internal/browser.js";
+} from "./browser.js";
 export {
   useClickOutside,
   useElementRef,
@@ -56,5 +85,5 @@ export {
   useFocusWithin,
   useHover,
   useIntersecting,
-} from "./internal/element.js";
-export { useCounter, useStorage, useToggle } from "./internal/state.js";
+} from "./dom.js";
+export { useCounter, useStorage, useToggle } from "./state.js";
